requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sentence-transformers==3.0.1
faiss-cpu==1.8.0
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.1
pydantic==2.8.2
pytest==8.3.2